# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tsllm",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/datasets": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Datasets",
        "operationId": "datasets_api_datasets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetOut"
                  },
                  "type": "array",
                  "title": "Response Datasets Api Datasets Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Dataset",
        "operationId": "dataset_api_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetailOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/datasets/{dataset_id}/ingest": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Ingest",
        "operationId": "ingest_api_datasets__dataset_id__ingest_post",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSubmittedOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/datasets/{dataset_id}/series": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Series",
        "description": "Bucket means over [start, end]; times are dataset-local and offsets are ignored.",
        "operationId": "series_api_datasets__dataset_id__series_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "channels",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeat for each channel.",
              "title": "Channels"
            },
            "description": "Repeat for each channel."
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "max_points",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20000,
              "minimum": 1,
              "default": 2000,
              "title": "Max Points"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/datasets/{dataset_id}/channels": {
      "put": {
        "tags": [
          "datasets"
        ],
        "summary": "Put Channels",
        "operationId": "put_channels_api_datasets__dataset_id__channels_put",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelsIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelsOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelSpec": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "Exact source column name."
          },
          "role": {
            "type": "string",
            "enum": [
              "target",
              "past_covariate",
              "known_future_covariate",
              "ignore"
            ],
            "title": "Role",
            "description": "Channel role.",
            "default": "target"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit",
            "description": "Physical unit."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Channel documentation."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ChannelSpec"
      },
      "ChannelsIn": {
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelSpec"
            },
            "type": "array",
            "minItems": 1,
            "title": "Channels",
            "description": "All channels; only role, unit, and description can change."
          }
        },
        "type": "object",
        "required": [
          "channels"
        ],
        "title": "ChannelsIn"
      },
      "ChannelsOut": {
        "properties": {
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          },
          "needs_ingest": {
            "type": "boolean",
            "title": "Needs Ingest"
          },
          "config": {
            "$ref": "#/components/schemas/DatasetConfig"
          }
        },
        "type": "object",
        "required": [
          "config_hash",
          "needs_ingest",
          "config"
        ],
        "title": "ChannelsOut"
      },
...
      "Condition": {
        "properties": {
          "channel": {
            "type": "string",
            "minLength": 1,
            "title": "Channel",
            "description": "Channel to compare."
          },
          "op": {
            "type": "string",
            "enum": [
              "gt",
              "ge",
              "lt",
              "le",
              "eq",
              "ne"
            ],
            "title": "Op",
            "description": "Comparison."
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Comparison threshold."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "channel",
          "op",
          "value"
        ],
        "title": "Condition"
      },
      "ConditionGroup": {
        "properties": {
          "all": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Condition"
                    },
                    {
                      "$ref": "#/components/schemas/ConditionGroup"
                    }
                  ]
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "All",
            "description": "All child rules must match."
          },
          "any": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Condition"
                    },
                    {
                      "$ref": "#/components/schemas/ConditionGroup"
                    }
                  ]
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Any",
            "description": "At least one child rule must match."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ConditionGroup"
      },
      "DatasetConfig": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]*$",
            "title": "Id",
            "description": "Dataset file stem."
          },
          "source": {
            "$ref": "#/components/schemas/SourceConfig",
            "description": "Source file configuration."
          },
          "native_freq": {
            "type": "string",
            "pattern": "^[0-9]+(?:[.][0-9]+)?(us|ms|s|min|h|d)$",
            "title": "Native Freq",
            "description": "Native sampling interval."
          },
          "freq": {
            "type": "string",
            "pattern": "^[0-9]+(?:[.][0-9]+)?(us|ms|s|min|h|d)$",
            "title": "Freq",
            "description": "Prepared grid interval."
          },
          "aggregation": {
            "type": "string",
            "enum": [
              "mean",
              "last"
            ],
            "title": "Aggregation",
            "description": "Bin aggregation.",
            "default": "mean"
          },
          "dedup": {
            "type": "string",
            "enum": [
              "keep_first",
              "keep_last"
            ],
            "title": "Dedup",
            "description": "Duplicate policy.",
            "default": "keep_last"
          },
          "channels": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChannelSpec"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels",
            "description": "Channels, or automatic numeric discovery."
          },
          "eligibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConditionGroup"
              },
              {
                "type": "null"
              }
            ],
            "description": "Eligibility rules, or any non-null channel."
          },
          "segments": {
            "$ref": "#/components/schemas/SegmentConfig",
            "description": "Continuous segment rules."
          },
          "split": {
            "$ref": "#/components/schemas/SplitConfig",
            "description": "Chronological split rules."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source",
          "native_freq",
          "freq",
          "segments"
        ],
        "title": "DatasetConfig"
      },
      "DatasetDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "fresh",
              "stale",
              "missing"
            ],
            "title": "Status",
            "description": "Cache state of the current configuration."
          },
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          },
          "config": {
            "$ref": "#/components/schemas/DatasetConfig"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta",
            "description": "Cache metadata; null unless fresh."
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "config_hash",
          "config",
          "meta"
        ],
        "title": "DatasetDetailOut"
      },
      "DatasetOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "fresh",
              "stale",
              "missing"
            ],
            "title": "Status",
            "description": "Cache state of the current configuration."
          },
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          },
          "config": {
            "$ref": "#/components/schemas/DatasetConfig"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "config_hash",
          "config"
        ],
        "title": "DatasetOut"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorOut"
      },
...
      "RunStatus": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled",
              "interrupted"
            ],
            "title": "State"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "state",
          "created_at"
        ],
        "title": "RunStatus"
      },
      "RunSubmittedOut": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunSubmittedOut"
      },
...
      "SegmentConfig": {
        "properties": {
          "max_gap": {
            "type": "string",
            "pattern": "^[0-9]+(?:[.][0-9]+)?(us|ms|s|min|h|d)$",
            "title": "Max Gap",
            "description": "Maximum internal interruption duration."
          },
          "min_length": {
            "type": "string",
            "pattern": "^[0-9]+(?:[.][0-9]+)?(us|ms|s|min|h|d)$",
            "title": "Min Length",
            "description": "Minimum segment duration, including grid steps."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "max_gap",
          "min_length"
        ],
        "title": "SegmentConfig"
      },
      "SegmentOut": {
        "properties": {
          "segment_id": {
            "type": "integer",
            "title": "Segment Id"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "length": {
            "type": "integer",
            "title": "Length"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          },
          "eligible_points": {
            "type": "integer",
            "title": "Eligible Points"
          }
        },
        "type": "object",
        "required": [
          "segment_id",
          "start",
          "end",
          "length",
          "duration_seconds",
          "eligible_points"
        ],
        "title": "SegmentOut"
      },
      "SeriesOut": {
        "properties": {
          "every": {
            "type": "string",
            "title": "Every",
            "description": "Bucket width; buckets are labelled with their start time."
          },
          "time": {
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "type": "array",
            "title": "Time"
          },
          "values": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "type": "object",
            "title": "Values"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentOut"
            },
            "type": "array",
            "title": "Segments"
          },
          "split_boundaries": {
            "additionalProperties": {
              "type": "string",
              "format": "date-time"
            },
            "type": "object",
            "title": "Split Boundaries"
          }
        },
        "type": "object",
        "required": [
          "every",
          "time",
          "values",
          "segments",
          "split_boundaries"
        ],
        "title": "SeriesOut"
      },
      "SourceConfig": {
        "properties": {
          "path": {
            "type": "string",
            "format": "path",
            "title": "Path",
            "description": "Source file path."
          },
          "format": {
            "type": "string",
            "enum": [
              "csv",
              "parquet"
            ],
            "title": "Format",
            "description": "Source format.",
            "default": "csv"
          },
          "encoding": {
            "type": "string",
            "enum": [
              "utf8",
              "utf8-lossy"
            ],
            "title": "Encoding",
            "description": "CSV encoding.",
            "default": "utf8"
          },
          "time_column": {
            "type": "string",
            "minLength": 1,
            "title": "Time Column",
            "description": "Timestamp column name."
          },
          "time_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Format",
            "description": "Timestamp format, or ISO discovery."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "path",
          "time_column"
        ],
        "title": "SourceConfig"
      },
      "SplitConfig": {
        "properties": {
          "ratios": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Ratios",
            "description": "Fit, val, cal, test proportions.",
            "default": [
              0.7,
              0.1,
              0.1,
              0.1
            ]
          },
          "basis": {
            "type": "string",
            "const": "eligible_points",
            "title": "Basis",
            "description": "Boundary count basis.",
            "default": "eligible_points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SplitConfig"
      },
...
    }
  }
}